+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Checkpoint] - 2026-01-05 07:51:39
+
+### Changed Files
+- `src/services/instagram/instagramService.ts`
+
+### Statistics
+```
+ 1 file changed, 20 insertions(+)
+```
+
 ## [Checkpoint] - 2026-01-04 18:17:47
 
 ### Changed Files